Temporary reduction of data chunks used for read.
We have a problem with our server that manifests itself when message exceeds read size (1024).
Message is continued in next read but our activemq returns part of next frame. In this case
frame is damaged - there is no cut in string nor explode by separator. This is temporary fix
(reducing read size to one character) until I find the time to investigate it properly.